Refactoring top level iotdevicecommon dir to avoid naming conflict.
