+++ devicedefender/tests/DeviceDefenderTest.cpp
@@ -4,7 +4,7 @@
  */
 #include <aws/crt/Api.h>
 
-#include <aws/iotdevice/IotDevice.h>
+#include <aws/iotdevicecommon/IotDevice.h>
 #include <aws/iotdevicedefender/DeviceDefender.h>
 #include <aws/testing/aws_test_harness.h>
 #include <utility>
@@ -14,7 +14,7 @@ static int s_TestDeviceDefenderResourceSafety(Aws::Crt::Allocator *allocator, vo
     (void)ctx;
     {
         Aws::Crt::ApiHandle apiHandle(allocator);
-        Aws::Iotdevice::DeviceApiHandle deviceApiHandle(allocator);
+        Aws::Iotdevicecommon::DeviceApiHandle deviceApiHandle(allocator);
         Aws::Crt::Io::TlsContextOptions tlsCtxOptions = Aws::Crt::Io::TlsContextOptions::InitDefaultClient();
 
         Aws::Crt::Io::TlsContext tlsContext(tlsCtxOptions, Aws::Crt::Io::TlsMode::CLIENT, allocator);
@@ -94,7 +94,7 @@ static int s_TestDeviceDefenderFailedTest(Aws::Crt::Allocator *allocator, void *
     (void)ctx;
     {
         Aws::Crt::ApiHandle apiHandle(allocator);
-        Aws::Iotdevice::DeviceApiHandle deviceApiHandle(allocator);
+        Aws::Iotdevicecommon::DeviceApiHandle deviceApiHandle(allocator);
         Aws::Crt::Io::TlsContextOptions tlsCtxOptions = Aws::Crt::Io::TlsContextOptions::InitDefaultClient();
 
         Aws::Crt::Io::TlsContext tlsContext(tlsCtxOptions, Aws::Crt::Io::TlsMode::CLIENT, allocator);
